# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Calculator API",
        "description": "API документация для приложения Calculator",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Основной сервер"
        }
    ],
    "paths": {
...
        "/news": {
            "get": {
                "tags": [
                    "News"
                ],
                "summary": "Список новостей",
                "operationId": "21b84dbdd75fd20b9c9ddb9b9e68f0a1",
                "responses": {
                    "200": {
                        "description": "Список новостей (HTML или JSON)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "News"
                ],
                "summary": "Создать новость",
                "operationId": "d6a12c2e3b4ea9ab588ba9b1cb14748f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "content"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "content": {
                                        "type": "string"
                                    },
                                    "images[]": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Перенаправление после создания"
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/news/{id}": {
            "get": {
                "tags": [
                    "News"
                ],
                "summary": "Детали новости",
                "operationId": "eb815321a4dd395639d4aeef25d9a45c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Данные новости"
                    },
                    "404": {
                        "description": "Новость не найдена"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "News"
                ],
                "summary": "Обновить новость",
                "operationId": "25ea273290c6cb379ab85e08ff4fadfc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "content"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "content": {
                                        "type": "string"
                                    },
                                    "images[]": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Перенаправление после обновления"
                    },
                    "404": {
                        "description": "Новость не найдена"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "News"
                ],
                "summary": "Удалить новость",
                "operationId": "bbe79a918bfbb58bc46ed74b43e9b986",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешное удаление (JSON)"
                    },
                    "302": {
                        "description": "Перенаправление после удаления"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/news/{id}/like": {
            "post": {
                "tags": [
                    "News"
                ],
                "summary": "Лайкнуть новость",
                "operationId": "f928560b4d8e83ad5ee230bc8a85e420",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Результат лайка (JSON)"
                    },
                    "302": {
                        "description": "Перенаправление"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/news/{newsId}/comments": {
            "post": {
                "tags": [
                    "News"
                ],
                "summary": "Добавить комментарий",
                "operationId": "bfd9cb95afb6d66d9e17b91364964f27",
                "parameters": [
                    {
                        "name": "newsId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "content"
                                ],
                                "properties": {
                                    "content": {
                                        "type": "string",
                                        "example": "Отличная новость!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Комментарий добавлен (JSON)"
                    },
                    "302": {
                        "description": "Перенаправление"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/pictures/{pictureId}": {
            "delete": {
                "tags": [
                    "News"
                ],
                "summary": "Удалить картинку новости",
                "operationId": "941b947c9968389bcf5e45b01ef6a4cb",
                "parameters": [
                    {
                        "name": "pictureId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешное удаление (JSON)"
                    },
                    "302": {
                        "description": "Перенаправление"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/news/{newsId}/comments/{commentId}": {
            "delete": {
                "tags": [
                    "News"
                ],
                "summary": "Удалить комментарий",
                "operationId": "0437fcad77d9095254a6375b7e9887cc",
                "parameters": [
                    {
                        "name": "newsId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "commentId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешное удаление (JSON)"
                    },
                    "302": {
                        "description": "Перенаправление"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}